requirements.txt
discord.py[voice]
pytube
opencv-python
numpy
requests